revert moving .storybook block and payload samples
