+++ src/.storybook/index.ts
@@ -1,3 +1,2 @@
 export * from './authHelpers'
-export * from './sampleBlock'
-export * from './samplePayload'
+export * from '../components/.storybook/samplePayload'
