--- src/contexts/Auth/authInterceptors/index.ts
@@ -1,1 +1,0 @@
-export * from './useAuthInterceptors'
